// src/app/Models/user-create.ts
export class UserCreate {
  accessDT: string = '';
  accessFeeder: string = '';
  accessOwner: string = '';
  accessSubdivision: string = '';
  accessSubStation: string = '';
  designation: string = '';
  email: string = '';
  roleID: string = '1';
  userName: string = '';
  writeAccess: string = '0';
  password: string = '';
  welcomemsg:string='';
  fromdate: string = '';
  todate: string = '';
}